tests: Fix a/z/c/z/test/zwave_command_class_time_parameters_test on UTC #46
This test was previously bypassed, it is now working as expected.

The reason it was failing is because mktime use timezone
to set the tm structure that will be returned to the next step of step.

Setting timezone to UTC (+0) solves the problem

Bug-SiliconLabsSoftware: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/issues/46
Origin: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/pull/69
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>