test: Add integration test using z-wave-stack-binaries to be used for CI
Check usage_ for how to use it.

This tool works along z-wave applications adapted for linux,
This package is maintained at z-wave-alliance and only available
to Z-Wave-Alliance members.

It is also supported in headless mode and can be run in a ci job,
as long are assets are made available (using GH_ZWAVE_ACCESS_TOKEN secret)

Alternatively it can be adapted to support real hardware,
the play recipe should not change.

Origin: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/pull/85
Relate-to: https://github.com/Z-Wave-Alliance/z-wave-stack-binaries/issues/5
Relate-to: https://github.com/Z-Wave-Alliance/z-wave-stack/issues/732
Relate-to: https://github.com/Z-Wave-Alliance/z-wave-stack-binaries/issues/2
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-engine-application-layer/issues/23
Relate-to: https://github.com/Z-Wave-Alliance/z-wave-stack/pull/734
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>