fix(zwapi) Harden zwapi_connection_tx in zwapi_connection.c
Change is obvious, it prevent an overflow,
Also it has been observed than on edge case (0xFF)
the counter never ends because it loops on the range of i
that was defined as a char, let's use a larger range.

note that functions did not test all inputs params
specially when it is done in caller
(eg: in  zwave_api_send_data reject frames above limit),
Also lenght is strored on 8bits which align to the frame max (0xFF/ 255).

Origin: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/pull/127
Bug-SiliconLabs: UIC-3666
Bug-SLVDBBP: 3169925
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>