fix(s2): Harden S2_init_ctx using safe memset
This is a simple change to check the static analysis.

SonarQube is reporting:

This "memset" is likely to be optimized away by the compiler; either remove it or replace it with "memset_s".

While memset_s is suggested it is not implemented widely (it is optional),
at this stage i am following GNU standards until it stabilizes.

One other possible change it to harden compilation flags with:

  -fno-builtin-memset

Origin: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/pull/137
Relate-to: https://www.open-std.org/jtc1/sc22/wg14/www/docs/n1967.htm
Relate-to: https://en.cppreference.com/w/c/string/byte/memset
Relate-to: https://sourceware.org/glibc/manual/2.41/html_node/Erasing-Sensitive-Data.html
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-protocol-controller/issues/100
Relate-to: https://github.com/SiliconLabsSoftware/z-wave-engine-application-layer/issues/42
Relate-to: https://sourceware.org/bugzilla/show_bug.cgi?id=17879#c2
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>